refactor: keep GitHub balance units bounded
