+++ apps/redskilled/src/daemon/lifecycle.ts
@@ -218,6 +218,7 @@ import {
 } from "./tunables.js";
 import { replaceWithViableSuccessor } from "./takeover.js";
 import { createRemotePollDeadline } from "./remote-poll.js";
+import { pollGithubBalanceObservers } from "./github-balance-observers.js";
 import { createConfiguredRedskilledSelfPingMonitor } from "./self-ping.js";
 import { resolveUnitDeath } from "./unit-death.js";
 // Error moved to ./daemon/errors.ts — keep re-export for backward compat
@@ -787,27 +788,15 @@ export async function startRedskilledDaemon(options: RedskilledDaemonOptions): P
         if (stored !== null) lastBalance = stored;
         return stored;
       }).finally(() => { balanceHydration = null; });
-      if ((await balanceHydration) !== null) {
-        await pollGithubBalanceObservers();
-        return lastBalance;
-      }
+      if ((await balanceHydration) !== null)
+        return pollGithubBalanceObservers(balanceRegistration.observers, remotePoll, clock).then(() => lastBalance);
     }
     lastBalance = await remotePoll("GitHub balance poll", () =>
       fetchGithubBalance({ transport: balanceRegistration.transport, now: clock() }));
     await balanceRegistration.store?.write(lastBalance).catch(() => undefined);
     await balanceRegistration.history?.append(lastBalance).catch(() => undefined);
-    await pollGithubBalanceObservers();
+    await pollGithubBalanceObservers(balanceRegistration.observers, remotePoll, clock);
     return lastBalance;
-  }
-
-  /** Observe optional payers without replacing the personal balance used by policy. */
-  async function pollGithubBalanceObservers(): Promise<void> {
-    for (const observer of balanceRegistration?.observers ?? []) {
-      const observed = await remotePoll(`GitHub balance poll (${observer.identity})`, () =>
-        fetchGithubBalance({ transport: observer.transport, now: clock() }));
-      await observer.store?.write(observed).catch(() => undefined);
-      await observer.history?.append(observed).catch(() => undefined);
-    }
   }
 
   /**
